+++ app/src/main/java/com/toyou/toyouandroid/presentation/fragment/onboarding/LoginViewModel.kt
@@ -1,14 +1,19 @@
 package com.toyou.toyouandroid.presentation.fragment.onboarding
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.toyou.toyouandroid.fcm.domain.FCMRepository
+import com.toyou.toyouandroid.fcm.dto.request.Token
 import com.toyou.toyouandroid.network.AuthNetworkModule
 import com.toyou.toyouandroid.presentation.fragment.onboarding.data.dto.request.SignUpRequest
 import com.toyou.toyouandroid.presentation.fragment.onboarding.data.dto.response.SignUpResponse
 import com.toyou.toyouandroid.presentation.fragment.onboarding.network.AuthService
 import com.toyou.toyouandroid.utils.TokenStorage
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import retrofit2.Call
 import retrofit2.Callback
@@ -19,6 +24,8 @@ class LoginViewModel(private val authService: AuthService, private val tokenStor
 
     private val _loginSuccess = MutableLiveData<Boolean>()
     val loginSuccess: LiveData<Boolean> get() = _loginSuccess
+    private val fcmRepository by lazy { FCMRepository(tokenStorage) }
+
 
     fun kakaoLogin(accessToken: String) {
         viewModelScope.launch {
@@ -142,5 +149,20 @@ class LoginViewModel(private val authService: AuthService, private val tokenStor
                 }
             })
         }
+    }
+
+    fun sendTokenToServer(token: String) {
+
+        CoroutineScope(Dispatchers.IO).launch {
+            try {
+                val tokenRequest = Token(token)
+                fcmRepository.postToken(tokenRequest)
+                Log.d("sendTokenToServer", "성공")
+            } catch (e: Exception) {
+                e.printStackTrace()
+                Log.d("sendTokenToServer", "토큰 전송 실패: ${e.message}")
+            }
+        }
+
     }
 }
