+++ app/src/main/res/layout/fragment_send_final.xml
@@ -19,7 +19,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="질문을 보냈습니다"
-        style="@style/gangwon15"
+        style="@style/gangwon25"
         android:layout_marginTop="30dp"
         app:layout_constraintTop_toBottomOf="@id/card_iv"
         app:layout_constraintEnd_toEndOf="parent"
